# public/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CNPJ Alfanumérico API",
    "description": "API completa para análise e correção de campos CNPJ em repositórios de código e bancos de dados. Sistema multi-perfil com gestão de tarefas, integrações com Azure DevOps/GitHub e chatbot inteligente.",
    "version": "2.0.0",
    "contact": {
      "name": "Suporte CNPJ Alfanumérico",
      "email": "suporte@cnpj-alfa.com"
    },
    "license": {
      "name": "Proprietário",
      "url": "https://cnpj-alfa.com/license"
    }
  },
  "servers": [
    {
      "url": "https://api.cnpj-alfa.com",
      "description": "Servidor de Produção"
    },
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento Local"
    }
  ],
...
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "tags": ["Autenticação"],
        "summary": "Obter usuário atual",
        "description": "Retorna informações do usuário autenticado atualmente",
        "operationId": "getCurrentUser",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/clients": {
      "get": {
        "tags": ["Super Admin"],
        "summary": "Listar todos os clientes",
        "description": "Retorna lista completa de clientes cadastrados (apenas Super Admin)",
        "operationId": "listClients",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 20}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "clients": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Client"}
                    },
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "totalPages": {"type": "integer"}
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": ["Super Admin"],
        "summary": "Criar novo cliente",
        "description": "Cria um novo cliente no sistema com licença e configurações",
        "operationId": "createClient",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "tags": ["Tarefas"],
        "summary": "Listar tarefas",
        "description": "Retorna lista de tarefas com filtros e paginação",
        "operationId": "listTasks",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pendente", "em_progresso", "concluida", "cancelada"]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["baixa", "media", "alta", "critica"]
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "schema": {"type": "string", "format": "uuid"}
          },
          {
            "name": "repository",
            "in": "query",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tarefas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Task"}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tarefas", "Admin Cliente"],
        "summary": "Criar nova tarefa",
        "description": "Cria uma tarefa manual de correção CNPJ",
        "operationId": "createTask",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}": {
      "put": {
        "tags": ["Tarefas"],
        "summary": "Atualizar tarefa",
        "description": "Atualiza informações de uma tarefa (status, prioridade, assignee)",
        "operationId": "updateTask",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarefa atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Tarefas", "Admin Cliente"],
        "summary": "Excluir tarefa",
        "description": "Remove uma tarefa do sistema",
        "operationId": "deleteTask",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {
          "204": {
            "description": "Tarefa excluída"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via endpoint /api/auth/login"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado - Token inválido ou ausente",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {"type": "string", "example": "Não autorizado"}
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Acesso negado - Permissão insuficiente",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {"type": "string", "example": "Acesso negado"}
              }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {"type": "string"},
                "details": {"type": "array", "items": {"type": "string"}}
              }
            }
          }
        }
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "email": {"type": "string", "format": "email"},
          "name": {"type": "string"},
          "role": {
            "type": "string",
            "enum": ["super_admin", "admin", "developer"]
          },
          "client_id": {"type": "string", "format": "uuid"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "Client": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "cnpj": {"type": "string"},
          "license_expires_at": {"type": "string", "format": "date-time"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateClientRequest": {
        "type": "object",
        "required": ["name", "cnpj", "admin_email", "admin_password"],
        "properties": {
          "name": {"type": "string", "example": "Empresa XYZ Ltda"},
          "cnpj": {"type": "string", "example": "12.345.678/0001-90"},
          "admin_email": {"type": "string", "format": "email"},
          "admin_password": {"type": "string", "format": "password"},
          "license_days": {"type": "integer", "default": 365}
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "analysis_id": {"type": "string", "format": "uuid"},
          "client_id": {"type": "string", "format": "uuid"},
          "repository": {"type": "string"},
          "file_path": {"type": "string"},
          "line_number": {"type": "integer"},
          "current_code": {"type": "string"},
          "suggested_code": {"type": "string"},
          "status": {
            "type": "string",
            "enum": ["pendente", "em_progresso", "concluida", "cancelada"]
          },
          "priority": {
            "type": "string",
            "enum": ["baixa", "media", "alta", "critica"]
          },
          "assignee_id": {"type": "string", "format": "uuid"},
          "external_id": {"type": "string", "description": "ID do Work Item no Azure DevOps"},
          "external_url": {"type": "string", "format": "uri"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateTaskRequest": {
        "type": "object",
        "required": ["repository", "file_path", "line_number", "current_code", "suggested_code"],
        "properties": {
          "repository": {"type": "string"},
          "file_path": {"type": "string"},
          "line_number": {"type": "integer"},
          "current_code": {"type": "string"},
          "suggested_code": {"type": "string"},
          "priority": {"type": "string", "default": "media"},
          "assignee_id": {"type": "string", "format": "uuid"}
        }
      },
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "priority": {"type": "string"},
          "assignee_id": {"type": "string", "format": "uuid"}
        }
      },
...
    }
  },
  "security": [
    {"BearerAuth": []}
  ]
}